convert component report configs
